pr05 Visual Regression Testing #1: Initial Visual Testing Framework (#1261)
* Update README.md

* added the image comparator which is the pixel matching algorithm

* added build.gradle file

* added the test runner

* added the simple test

* Revise README for Jetpack Compose migration strategy

Updated README to reflect migration to Jetpack Compose and strategy for replacing JEditTextArea with RSyntaxTextArea. Added insights on LSP-based editor research and the need for user feedback on Tweak Mode and autocompletion features.

* fixing the build issues

* added junit as dependency

* removing custom class implementation

* inclding visual-tests in settings

* fixed the overlapping cmd

* cleaning

* adding packages

* added updated screenshot structure

* refactoring

* added tests in suits

* removed simple test

* deleting earlier files

* updated the core/gradle file

* added the infrastructure

* added some tests ported by p5js

* removing test rendering suite and its test file

* added screenshots

* config files

* fixed the pixeldensity to 1

* Revert "fixed the pixeldensity to 1"

This reverts commit 66071ac191d24a4a25d8d8257f58cbeb957c6f35.

* fixed pixeldensity to 1

* Configure dependencyUpdates task in build.gradle.kts

Add configuration for dependencyUpdates task to manage non-stable versions.

* removing rendering gradient screenshot

* General cleanup of `Base`

I started cleaning up some of `Base`'s startup sequence for clarity of what is being started when. Nowhere near completion and I think a lot of this class will need to be refactored in the future.

Also removed some of the timing measurement comments

Added some comments to the Processing CLI class

* Move contributor list to CONTRIBUTORS.md (#1313)

Created CONTRIBUTORS.md and updated .all-contributorsrc to reference the new file instead of README.md. This will reduce the size of the README and improve loading times.

* Update BUILD.md with build failure troubleshooting

Added troubleshooting steps for build failures related to permissions.

* fixing the build issues

* inclding visual-tests in settings

* updated the core/gradle file

* config files

* Configure dependencyUpdates task in build.gradle.kts

Add configuration for dependencyUpdates task to manage non-stable versions.

* fix rebasing

---------

Co-authored-by: Stef Tervelde <stef@steftervelde.nl>
Co-authored-by: Raphaël de Courville <raphael@processingfoundation.org>